refactor backend components
